Querying Objects

+++ Vidly/Models/Customer.cs
@@ -6,14 +6,13 @@ using System.Web;
 
 namespace Vidly.Models
 {
-    public class Customer
-    {
+    public class Customer{
         public int Id { get; set; }
         [Required]
         [StringLength(255)]
         public string Name { get; set; }
         public bool IsSubscribedToNewsLetter { get; set; }
         public MembershipType MembershipType { get; set; }
-        public byte MembershipTypeId { get; set; }
+        public short MembershipTypeId { get; set; }
     }
 }
